requirements.txt
# Core
python-dotenv==1.0.0
requests==2.31.0

# Apify
apify-client==1.7.1

# Database
sqlalchemy==2.0.23

# Dashboard
dash==2.14.2
dash-bootstrap-components==1.5.0
plotly==5.18.0
pandas==2.1.4

# Utils
python-dateutil==2.8.2
pytz==2023.3
